ctrl d fixed when used at heredoc, get_next_token functions
