+++ business/lexer.c
@@ -6,7 +6,7 @@
 /*   By: aderviso <aderviso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 18:32:26 by aderviso          #+#    #+#             */
-/*   Updated: 2023/06/12 19:30:32 by aderviso         ###   ########.fr       */
+/*   Updated: 2023/06/13 15:13:04 by aderviso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,11 +77,14 @@ void	get_next_token(t_token *command_table)
 	while (tmp_command_table->next)
 		tmp_command_table = tmp_command_table->next;
 	input = readline("> ");
-	ctrl_d(str);
+	ctrl_d_as_eof(input);
 	if (g_ms.ignore)
-		exit(1);
-	new_tokens = tokenizer(input);
+		print_error(258, "end of file");
+	else
+	{
+		new_tokens = tokenizer(input);
+		new_tokens->prev = tmp_command_table;
+		tmp_command_table->next = new_tokens;		
+	}
 	free(input);
-	new_tokens->prev = tmp_command_table;
-	tmp_command_table->next = new_tokens;
 }
